Move helper test methods to base class
