+++ test/OpenTelemetry.Tests/Metrics/MetricAPITest.cs
@@ -27,7 +27,7 @@ using Xunit.Abstractions;
 
 namespace OpenTelemetry.Metrics.Tests
 {
-    public class MetricApiTest
+    public class MetricApiTest : MetricTestsBase
     {
         private const int MaxTimeToAllowForFlush = 10000;
         private static readonly int NumberOfThreads = Environment.ProcessorCount;
@@ -1547,109 +1547,6 @@ namespace OpenTelemetry.Metrics.Tests
 
             meterProvider.ForceFlush(MaxTimeToAllowForFlush);
             Assert.Empty(exportedItems);
-        }
-
-        private static void ValidateMetricPointTags(List<KeyValuePair<string, object>> expectedTags, ReadOnlyTagCollection actualTags)
-        {
-            int tagIndex = 0;
-            foreach (var tag in actualTags)
-            {
-                Assert.Equal(expectedTags[tagIndex].Key, tag.Key);
-                Assert.Equal(expectedTags[tagIndex].Value, tag.Value);
-                tagIndex++;
-            }
-
-            Assert.Equal(expectedTags.Count, tagIndex);
-        }
-
-        private static long GetLongSum(List<Metric> metrics)
-        {
-            long sum = 0;
-            foreach (var metric in metrics)
-            {
-                foreach (ref readonly var metricPoint in metric.GetMetricPoints())
-                {
-                    if (metric.MetricType.IsSum())
-                    {
-                        sum += metricPoint.GetSumLong();
-                    }
-                    else
-                    {
-                        sum += metricPoint.GetGaugeLastValueLong();
-                    }
-                }
-            }
-
-            return sum;
-        }
-
-        private static double GetDoubleSum(List<Metric> metrics)
-        {
-            double sum = 0;
-            foreach (var metric in metrics)
-            {
-                foreach (ref readonly var metricPoint in metric.GetMetricPoints())
-                {
-                    if (metric.MetricType.IsSum())
-                    {
-                        sum += metricPoint.GetSumDouble();
-                    }
-                    else
-                    {
-                        sum += metricPoint.GetGaugeLastValueDouble();
-                    }
-                }
-            }
-
-            return sum;
-        }
-
-        private static int GetNumberOfMetricPoints(List<Metric> metrics)
-        {
-            int count = 0;
-            foreach (var metric in metrics)
-            {
-                foreach (ref readonly var metricPoint in metric.GetMetricPoints())
-                {
-                    count++;
-                }
-            }
-
-            return count;
-        }
-
-        private static MetricPoint? GetFirstMetricPoint(List<Metric> metrics)
-        {
-            foreach (var metric in metrics)
-            {
-                foreach (ref readonly var metricPoint in metric.GetMetricPoints())
-                {
-                    return metricPoint;
-                }
-            }
-
-            return null;
-        }
-
-        // Provide tags input sorted by Key
-        private static void CheckTagsForNthMetricPoint(List<Metric> metrics, List<KeyValuePair<string, object>> tags, int n)
-        {
-            var metric = metrics[0];
-            var metricPointEnumerator = metric.GetMetricPoints().GetEnumerator();
-
-            for (int i = 0; i < n; i++)
-            {
-                Assert.True(metricPointEnumerator.MoveNext());
-            }
-
-            int index = 0;
-            var metricPoint = metricPointEnumerator.Current;
-            foreach (var tag in metricPoint.Tags)
-            {
-                Assert.Equal(tags[index].Key, tag.Key);
-                Assert.Equal(tags[index].Value, tag.Value);
-                index++;
-            }
         }
 
         private static void CounterUpdateThread<T>(object obj)
